backend/requirements.txt
Django==5.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyJWT==2.8.0
Pillow==10.1.0